refactor: rename helper package to aggregates, implement password hashing, and update user creation logic
