+++ identification/pkg/helper/hashing.go
@@ -1,0 +1,12 @@
+package helper
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+)
+
+func HashPassword(password string) string {
+	hasher := sha256.New()
+	hasher.Write([]byte(password))
+	return hex.EncodeToString(hasher.Sum(nil))
+}
