+++ identification/pkg/models/query/token_request.go
@@ -1,5 +1,5 @@
 package query
 
 type TokenRequest struct {
-	PasswordHash string `json:"password_hash"`
+	Password string `json:"password"`
 }
