+++ identification/pkg/query/eventpolling/polling.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/L4B0MB4/EVTSRC/pkg/client"
 	"github.com/L4B0MB4/PRYVT/identification/pkg/aggregates"
-	"github.com/L4B0MB4/PRYVT/identification/pkg/helper"
 	"github.com/L4B0MB4/PRYVT/identification/pkg/query/store/repository"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
@@ -46,7 +45,7 @@ func (ep *EventPolling) PollEvents() {
 					log.Err(err).Msg("Error while creating user aggregate")
 					break
 				}
-				uI := helper.GetUserModelFromAggregate(ua)
+				uI := aggregates.GetUserModelFromAggregate(ua)
 				err = ep.userRepo.AddOrReplaceUser(uI)
 				if err != nil {
 					log.Err(err).Msg("Error while adding or replacing user")
