+++ identification/pkg/query/httphandler/controller/user_controller.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 
 	"github.com/L4B0MB4/PRYVT/identification/pkg/aggregates"
+	"github.com/L4B0MB4/PRYVT/identification/pkg/helper"
 	models "github.com/L4B0MB4/PRYVT/identification/pkg/models/query"
 	"github.com/L4B0MB4/PRYVT/identification/pkg/query/authentication"
 	"github.com/L4B0MB4/PRYVT/identification/pkg/query/store/repository"
@@ -37,7 +38,8 @@ func (ctrl *UserController) GetToken(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "User not found"})
 		return
 	}
-	if user.PasswordHash != tokenReq.PasswordHash {
+	hashedPw := helper.HashPassword(tokenReq.Password)
+	if user.PasswordHash != hashedPw {
 
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
 		return
